changing to targets/targets to confirm preds are problem

+++ ultralytics/models/yolo/detect/val.py
@@ -103,7 +103,7 @@ class SequenceValidator():
                     # Only assign if box is in correct format
                     if((x1 < x2) and (y1 < y2)):
                         pred_boxes.append(torch.tensor([x1, y1, x2, y2]).to(self.device))
-                        pred_cls.append(torch.tensor(cls).to(self.device))
+                        pred_cls.append(torch.tensor(cls).to(torch.int).to(self.device))
                         pred_scores.append(torch.tensor(score).to(self.device))
                 # Filter no detection and single detection
                 #boxes = torch.clip(torch.stack(pred_boxes, dim=0), min=0, max=1280)
@@ -118,7 +118,7 @@ class SequenceValidator():
                     #labels = torch.tensor([])'''
                 preds.append({
                     'boxes': torch.clip(torch.stack(pred_boxes, dim=0), min=0, max=1280),
-                    'labels':torch.stack(pred_cls, dim=0).to(torch.int),
+                    'labels':torch.stack(pred_cls, dim=0),
                     'scores':torch.stack(pred_scores)
                 })
             #print(f"Targets({len(targets)}): boxes-{targets[0]['boxes'].shape}, labels-{targets[0]['labels']}")
